simutil_verilator: VerilatedToplevel is now "top"
Previously, we had the following class hierarchy:

                      ibex_riscv_compliance
                      /                   \
          Vibex_riscv_compliance    VerilatedToplevel

Vibex_riscv_compliance is auto-generated by Verilator, and ibex_riscv_compliance
is generated by the VERILATED_TOPLEVEL() macro, and VerilatedToplevel is
an abstract class.

This diamond-shaped inheritance hierarchy makes it possible to
type-annotate arguments that need access to eval() or finish() with the generic
VerilatedToplevel class type.

After this change, we have a simpler hierarchy:

                        VerilatedToplevel
                                |
                      Vibex_riscv_compliance

VerilatedToplevel is macro-generated from the |toplevel| define.