[verilator] Support memory initialization from ELF file
Load ELF segments into memory structures for Verilator simulations.

The content of the ELF segments are stored successively in an internal
buffer before setting the memory. Only loadable segments are included.

The memory must provide `simutil_verilator_set_mem` as a way to set the
content.

Loading of vmem files is still possible. The file extension is used to
distinguish between the different file formats.