[verilator] Change memory init argument interface
Remove the explicit arguments for setting a specific memory section.
Introduce a new argument which allows naming a memory section for
initialization. The argument can be used multiple times in order to
initialize multiple memory regions. A special keyword 'list' is used to
print the registered memory regions.

The available memory region must be specified by using
`RegisterMemoryArea()` and are identified by a name which must be
supplied by the memory initialization argument. The memory must be
registered before it can be used by `ParseCommandArgs`.Setting the
type of the file is optional. If not set, determination is based on the
file extension.

The exit code is set more explicitly based on the error type.