[verilator] Ignore empty memory size
A segment can have a memory size which is larger that the file size.
This space should then be made available when the file is loaded.
The ELF loading for the simulation can not place the sections at the
virtual memory address as it has no notion of the systems addresses.
Ignore those parts of the ELF file. If a initialization of a region is
necessary, this can be achieved by setting symbols while running the
linker and a related startup sequence.